voice-secretary/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
googletrans==4.0.0rc1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1